### 2024-04-05 #6585 @arfon
👋 @xianggebenben, thanks for your submission to JOSS. Can you confirm which aspects of the repository are code that should be reviewed here? It looks like you've packaged some dependencies in the repository which I'm assuming should *not* be considered part of the submission/your direct work.
