### 2024-04-25 #6585 @majensen
@range-et - [this work](#6585) just came in to JOSS - would you be interested in reviewing it? Thanks for considering! --@majensen
