+++ README.md
@@ -1,6 +1,8 @@
 [![Documentation Status](https://readthedocs.org/projects/graphsl/badge/?version=latest)](https://graphsl.readthedocs.io/en/latest/?badge=latest)
 [![MIT license](https://img.shields.io/badge/License-MIT-blue.svg)](https://lbesson.mit-license.org/)
 [![PyPI version](https://badge.fury.io/py/graphsl.svg)](https://badge.fury.io/py/graphsl)
+[![Downloads](https://pepy.tech/badge/graphsl)](https://pepy.tech/project/graphsl)
+
 
 # GraphSL: Graph Source Localization Library
 
