+++ GraphSL/GNN/GCNSI/main.py
@@ -54,7 +54,7 @@ class GCNSI:
 
         Example:
 
-        from GraphSL.data.utils import load_dataset, diffusion_generation, split_dataset
+        from GraphSL.utils import load_dataset, diffusion_generation, split_dataset
 
         from GraphSL.GNN.GCNSI.main import GCNSI
 
