+++ GraphSL/__init__.py
@@ -1,2 +1,3 @@
 from .Evaluation import *
 from .Prescribed import *
+from .utils import *
