+++ GraphSL.egg-info/SOURCES.txt
@@ -1,31 +1,0 @@
-LICENSE
-README.md
-setup.cfg
-setup.py
-GraphSL/Evaluation.py
-GraphSL/Prescribed.py
-GraphSL/__init__.py
-GraphSL.egg-info/PKG-INFO
-GraphSL.egg-info/SOURCES.txt
-GraphSL.egg-info/dependency_links.txt
-GraphSL.egg-info/requires.txt
-GraphSL.egg-info/top_level.txt
-GraphSL/GNN/__init__.py
-GraphSL/GNN/GCNSI/__init__.py
-GraphSL/GNN/GCNSI/main.py
-GraphSL/GNN/GCNSI/model.py
-GraphSL/GNN/IVGD/__init__.py
-GraphSL/GNN/IVGD/correction.py
-GraphSL/GNN/IVGD/earlystopping.py
-GraphSL/GNN/IVGD/i_deepis.py
-GraphSL/GNN/IVGD/main.py
-GraphSL/GNN/IVGD/preprocessing.py
-GraphSL/GNN/IVGD/training.py
-GraphSL/GNN/IVGD/validity_net.py
-GraphSL/GNN/IVGD/model/MLP.py
-GraphSL/GNN/IVGD/model/__init__.py
-GraphSL/GNN/SLVAE/__init__.py
-GraphSL/GNN/SLVAE/main.py
-GraphSL/GNN/SLVAE/model.py
-GraphSL/data/__init__.py
-GraphSL/data/utils.py
