+++ GraphSL/GNN/GCNSI/main.py
@@ -54,13 +54,17 @@ class GCNSI:
 
         Example:
 
+        import os
+
+        curr_dir = os.getcwd()
+
         from GraphSL.utils import load_dataset, diffusion_generation, split_dataset
 
         from GraphSL.GNN.GCNSI.main import GCNSI
 
         data_name = 'karate'
 
-        graph = load_dataset(data_name)
+        graph = load_dataset(data_name, data_dir=curr_dir)
 
         dataset = diffusion_generation(graph=graph, infect_prob=0.3, diff_type='IC', sim_num=100, seed_ratio=0.1)
 
@@ -183,13 +187,17 @@ class GCNSI:
 
         Example:
 
+        import os
+
+        curr_dir = os.getcwd()
+
         from data.utils import load_dataset, diffusion_generation, split_dataset
 
         from GNN.GCNSI.main import GCNSI
 
         data_name = 'karate'
 
-        graph = load_dataset(data_name)
+        graph = load_dataset(data_name, data_dir=curr_dir)
 
         dataset = diffusion_generation(graph=graph, infect_prob=0.3, diff_type='IC', sim_num=100, seed_ratio=0.1)
 
