+++ GraphSL/GNN/SLVAE/main.py
@@ -194,13 +194,17 @@ class SLVAE:
 
         Example:
 
+        import os
+
+        curr_dir = os.getcwd()
+
         from GraphSL.utils import load_dataset, diffusion_generation, split_dataset
 
         from GraphSL.GNN.SLVAE.main import SLVAE
 
         data_name = 'karate'
 
-        graph = load_dataset(data_name)
+        graph = load_dataset(data_name, data_dir=curr_dir)
 
         dataset = diffusion_generation(graph=graph, infect_prob=0.3, diff_type='IC', sim_num=100, seed_ratio=0.1)
 
@@ -357,13 +361,17 @@ class SLVAE:
 
         Example:
 
+        import os
+
+        curr_dir = os.getcwd()
+
         from GraphSL.utils import load_dataset, diffusion_generation, split_dataset
 
         from GraphSL.GNN.SLVAE.main import SLVAE
 
         data_name = 'karate'
 
-        graph = load_dataset(data_name)
+        graph = load_dataset(data_name, data_dir=curr_dir)
 
         dataset = diffusion_generation(graph=graph, infect_prob=0.3, diff_type='IC', sim_num=100, seed_ratio=0.1)
 
