+++ GraphSL/GNN/IVGD/main.py
@@ -148,13 +148,17 @@ class IVGD:
 
         Example:
 
+        import os
+
+        curr_dir = os.getcwd()
+
         from GraphSL.utils import load_dataset, diffusion_generation, split_dataset
 
         from GraphSL.GNN.IVGD.main import IVGD
 
         data_name = 'karate'
 
-        graph = load_dataset(data_name)
+        graph = load_dataset(data_name, data_dir=curr_dir)
 
         dataset = diffusion_generation(graph=graph, infect_prob=0.3, diff_type='IC', sim_num=100, seed_ratio=0.1)
 
@@ -237,13 +241,17 @@ class IVGD:
 
         Example:
 
+        import os
+
+        curr_dir = os.getcwd()
+
         from GraphSL.utils import load_dataset, diffusion_generation, split_dataset
 
         from GraphSL.GNN.IVGD.main import IVGD
 
         data_name = 'karate'
 
-        graph = load_dataset(data_name)
+        graph = load_dataset(data_name, data_dir=curr_dir)
 
         dataset = diffusion_generation(graph=graph, infect_prob=0.3, diff_type='IC', sim_num=100, seed_ratio=0.1)
 
@@ -352,13 +360,17 @@ class IVGD:
 
         Example:
 
+        import os
+
+        curr_dir = os.getcwd()
+
         from GraphSL.utils import load_dataset, diffusion_generation, split_dataset
 
         from GraphSL.GNN.IVGD.main import IVGD
 
         data_name = 'karate'
 
-        graph = load_dataset(data_name)
+        graph = load_dataset(data_name, data_dir=curr_dir)
 
         dataset = diffusion_generation(graph=graph, infect_prob=0.3, diff_type='IC', sim_num=100, seed_ratio=0.1)
 
