+++ build/lib/GraphSL/GNN/SLVAE/model.py
@@ -1,0 +1,466 @@
+import torch.nn as nn
+import torch
+import torch.nn.functional as F
+import scipy.sparse as sp
+class Encoder(nn.Module):
+    """
+    Encoder module for a variational autoencoder (VAE).
+
+    Attributes:
+
+    - input_dim (int): Dimension of the input.
+
+    - hidden_dim (int): Dimension of the hidden layer.
+
+    - latent_dim (int): Dimension of the latent space.
+    """
+
+    def __init__(self, input_dim=784, hidden_dim=512, latent_dim=256):
+        """
+        Initialize the Encoder.
+
+        Args:
+
+        - input_dim (int): Dimension of the input.
+
+        - hidden_dim (int): Dimension of the hidden layer.
+
+        - latent_dim (int): Dimension of the latent space.
+        """
+        super(Encoder, self).__init__()
+
+        self.linear1 = nn.Linear(input_dim, hidden_dim)
+        self.linear2 = nn.Linear(hidden_dim, hidden_dim)
+        self.mean = nn.Linear(hidden_dim, latent_dim)
+        self.var = nn.Linear(hidden_dim, latent_dim)
+        self.LeakyReLU = nn.LeakyReLU(0.2)
+
+    def forward(self, x):
+        """
+        Forward pass of the Encoder.
+
+        Args:
+        - x (torch.Tensor): Input tensor.
+
+        Returns:
+
+        - mean (torch.Tensor): The mean of the latent space.
+
+        - log_var (torch.Tensor): The log variance of the latent space.
+
+        """
+        x = self.LeakyReLU(self.linear1(x))
+        x = self.LeakyReLU(self.linear2(x))
+
+        mean = self.mean(x)
+        log_var = self.var(x)
+        return mean, log_var
+
+
+class Decoder(nn.Module):
+    """
+    Decoder module for a variational autoencoder (VAE).
+
+    Attributes:
+
+    - output_dim (int): Dimension of the output.
+
+    - hidden_dim (int): Dimension of the hidden layer.
+
+    - latent_dim (int): Dimension of the latent space.
+    """
+
+    def __init__(self, output_dim=784, hidden_dim=512, latent_dim=256):
+        """
+        Initialize the Decoder.
+
+        Args:
+
+        - output_dim (int): Dimension of the output.
+
+
+        - hidden_dim (int): Dimension of the hidden layer.
+
+        - latent_dim (int): Dimension of the latent space.
+        """
+        super(Decoder, self).__init__()
+
+        self.linear2 = nn.Linear(latent_dim, hidden_dim)
+        self.linear1 = nn.Linear(hidden_dim, hidden_dim)
+        self.output = nn.Linear(hidden_dim, output_dim)
+        self.LeakyReLU = nn.LeakyReLU(0.2)
+
+    def forward(self, x):
+        """
+        Forward pass of the Decoder.
+
+        Args:
+
+        - x (torch.Tensor): Input tensor.
+
+        Returns:
+
+        - x_hat (torch.Tensor): Decoded output tensor.
+        """
+        x = self.LeakyReLU(self.linear2(x))
+        x = self.LeakyReLU(self.linear1(x))
+
+        x_hat = torch.sigmoid(self.output(x))
+        return x_hat
+
+
+
+class VAE(nn.Module):
+    """
+    Variational Autoencoder (VAE) model.
+
+    Attributes:
+
+    - input_dim (int): Dimension of the input.
+
+    - hidden_dim (int): Dimension of the hidden layer.
+
+    - latent_dim (int): Dimension of the latent space.
+    """
+
+    def __init__(self, input_dim=1, hidden_dim=512, latent_dim=256):
+        """
[…]
