+++ build/lib/GraphSL/GNN/IVGD/i_deepis.py
@@ -1,0 +1,198 @@
+import torch
+import torch.nn as nn
+import numpy as np
+import scipy.sparse as sp
+
+
+class Identity(nn.Module):
+    """
+    Identity module to select specific elements from input tensor.
+    """
+
+    def __init__(self):
+        super(Identity, self).__init__()
+
+    def forward(self, preds, idx):
+        """
+        Forward pass to select specific elements from input tensor.
+
+        Args:
+
+        - preds (torch.Tensor): Input tensor.
+
+        - idx (torch.Tensor): Indices of elements to select.
+
+        Returns:
+
+        - torch.Tensor: Selected elements from input tensor.
+        """
+        return preds[idx]
+
+
+class DiffusionPropagate(nn.Module):
+    """
+    Module for diffusion propagation.
+    """
+
+    def __init__(self, prob_matrix, niter):
+        """
+        Initializes the DiffusionPropagate module.
+
+        Args:
+
+        - prob_matrix (torch.Tensor): Probability matrix for diffusion.
+
+        - niter (int): Number of iterations for diffusion propagation.
+        """
+        super(DiffusionPropagate, self).__init__()
+        self.niter = niter
+        if sp.isspmatrix(prob_matrix):
+            prob_matrix = prob_matrix.toarray()
+        self.register_buffer('prob_matrix', torch.FloatTensor(prob_matrix))
+
+    def forward(self, preds, seed_idx, idx):
+        """
+        Performs forward pass for diffusion propagation.
+
+        Args:
+
+        - preds (torch.Tensor): Input tensor of predictions.
+
+        - seed_idx (torch.Tensor): Indices of seed nodes.
+
+        - idx (torch.Tensor): Indices of nodes to propagate to.
+
+        Returns:
+
+        - torch.Tensor: Resulting propagated predictions.
+        """
+        temp = preds
+        temp = temp.flatten()
+        device = preds.device
+        for i in range(self.niter):
+            P2 = self.prob_matrix.T * preds.view((1, -1)).expand(self.prob_matrix.shape)
+            P3 = torch.ones(self.prob_matrix.shape).to(device) - P2
+            preds = torch.ones((self.prob_matrix.shape[0],)).to(device) - torch.prod(P3, dim=1)
+            #preds[seed_idx] = 1
+        preds = (preds + temp) / 2
+        return preds[idx]
+
+    def backward(self, preds):
+        """
+        Perform backward pass for diffusion propagation.
+
+        Args:
+
+        - preds (torch.Tensor):  Prediction of diffusion.
+
+        Returns:
+
+        - res (torch.Tensor): Prediction of seeds.
+        """
+        device = preds.device
+        res = preds
+        temp = preds
+        for j in range(10):
+            for i in range(self.niter):
+                P2 = self.prob_matrix.T * res.view((1, -1)).expand(self.prob_matrix.shape)
+                P3 = torch.ones(self.prob_matrix.shape).to(device) - P2
+                temp = torch.ones((self.prob_matrix.shape[0],)).to(device) - torch.prod(P3, dim=1)
+                #temp[preds == 1] = 1
+            res = 2 * preds - temp
+            res = torch.maximum(torch.minimum(res, torch.tensor(1)), torch.tensor(0))
+        return res
+
+
+class i_DeepIS(nn.Module):
+    """
+    Invertible Deep Influence Spread (i_DeepIS) module for graph source localization.
+    """
+
+    def __init__(self, gnn_model: nn.Module, propagate: nn.Module):
+        """
+        Initializes the i_DeepIS module.
+
+        Args:
+
+        - gnn_model (nn.Module): Graph neural network model.
+
+        - propagate (nn.Module): Propagation module for diffusion.
[…]
