+++ build/lib/GraphSL/GNN/SLVAE/__init__.py
@@ -1,0 +1,2 @@
+from .main import *
+from .model import *
