+++ paper.bib
@@ -28,7 +28,7 @@
 
 @inproceedings{IVGD_www22,
 title={An Invertible Graph Diffusion Neural Network for Source Localization},
-author={Junxiang Wang, Junji Jiang, and Liang Zhao},
+author={Wang, Junxiang and Jiang, Junji and Zhao, Liang},
 booktitle={Proceedings of the 31th International World Wide Web Conference (WWW 2022)},
 year={2022}
 
@@ -86,7 +86,7 @@ year={2022}
 
 @article{frkaszczak2022rpasdt,
   title={RPaSDT—rumor propagation and source detection Toolkit},
-  author={Fr{\k{a}}szczak, Damian},
+  author={Fraszczak, Damian},
   journal={SoftwareX},
   volume={17},
   pages={100988},
