+++ README.md
@@ -146,7 +146,7 @@ If you use this package in your research, please consider citing our work as fol
   year = {2024},
   author = {Wang Junxiang, Zhao Liang},
   title = {GraphSL: A Open-Source Library for Graph Source Localization Approaches and Benchmark Datasets},
-  journal = {preprint}
+  journal = {preprint, 	arXiv:2405.03724}
 }
 ```
 # Contact
