+++ requirements.txt
@@ -76,7 +76,7 @@ pytz==2024.1
 PyYAML==6.0.1
 pyzmq==25.1.2
 referencing==0.34.0
-requests==2.31.0
+requests==2.32.0
 rpds-py==0.18.0
 scikit-learn==1.4.1.post1
 scipy==1.12.0
