### 2024-05-28 #6796 @editorialbot
[REVIEW]: GraphSL: A Open-Source Library for Graph Source Localization Approaches and Benchmark Datasets
**Submitting author:** @xianggebenben (<a href="http://orcid.org/0000-0002-6635-4296">Junxiang Wang</a>)
**Repository:** https://github.com/xianggebenben/GraphSL
**Branch with paper.md** (empty if default branch):
**Version:** 0.15
