### 2024-05-28 xianggebenben/GraphSL#7 @mbeyss
format according to pep8
The code contains many violation of pep8, which make the code harder to read. Formatting should ideally be done automatically (e.g. black, flake8, autopep8), and enforced on commits (pre-commit).

related to this [JOSS review](https://github.com/openjournals/joss-reviews/issues/6796)
