+++ GraphSL/GNN/IVGD/training.py
@@ -1,4 +1,6 @@
-from typing import  Tuple
+from .earlystopping import EarlyStopping, stopping_args
+from .preprocessing import gen_seeds, gen_splits_
+from typing import Tuple
 import time
 import numpy as np
 import torch
@@ -7,8 +9,7 @@ from torch.utils.data import TensorDataset, DataLoader
 import scipy.sparse as sp
 import sys
 sys.path.append('../../')
-from .preprocessing import gen_seeds, gen_splits_
-from .earlystopping import EarlyStopping, stopping_args
+
 
 class FeatureCons:
     """
@@ -84,9 +85,17 @@ def get_dataloaders(idx, labels_np, batch_size=None):
     labels = torch.FloatTensor(labels_np)
     if batch_size is None:
         batch_size = max((val.numel() for val in idx.values()))
-    datasets = {phase: TensorDataset(torch.LongTensor(ind), labels[ind]) for phase, ind in idx.items()}
-    dataloaders = {phase: DataLoader(dataset=dataset, batch_size=batch_size, shuffle=True) for phase, dataset in
-                   datasets.items()}
+    datasets = {
+        phase: TensorDataset(
+            torch.LongTensor(ind),
+            labels[ind]) for phase,
+        ind in idx.items()}
+    dataloaders = {
+        phase: DataLoader(
+            dataset=dataset,
+            batch_size=batch_size,
+            shuffle=True) for phase,
+        dataset in datasets.items()}
     return dataloaders
 
 
@@ -143,7 +152,8 @@ def lanczos_algo(prob_matrix, seed_vec, order=5, epsilon=0.001):
     seed_vec = seed_vec.flatten()
     beta = np.zeros((order + 1,))
     gamma = np.zeros((order + 1,))
-    q = [np.zeros((len(seed_vec),)), seed_vec / np.linalg.norm(seed_vec, ord=2)]
+    q = [np.zeros((len(seed_vec),)), seed_vec /
+         np.linalg.norm(seed_vec, ord=2)]
     for j in range(1, order + 1):
         z = S @ q[j]
         gamma[j] = q[j].reshape((1, -1)) @ z
@@ -172,16 +182,20 @@ def update_embedding(model, feature_mat):
 
     - torch.nn.Module: Updated model.
     """
-    assert getattr(model, 'gnn_model', None) is not None, 'Object model should have a submodule `gnn_model` '
+    assert getattr(model, 'gnn_model',
+                   None) is not None, 'Object model should have a submodule `gnn_model` '
     device = next(model.parameters()).device
     if model.gnn_model.features is None:
-        new_embedding_layer = nn.Embedding(feature_mat.shape[0], feature_mat.shape[1])
-        new_embedding_layer.weight = nn.Parameter(torch.FloatTensor(feature_mat), requires_grad=False)
+        new_embedding_layer = nn.Embedding(
+            feature_mat.shape[0], feature_mat.shape[1])
+        new_embedding_layer.weight = nn.Parameter(
+            torch.FloatTensor(feature_mat), requires_grad=False)
         model.gnn_model.features = new_embedding_layer
     else:
         assert feature_mat.shape[1] == model.gnn_model.features.weight.shape[1], \
             'New dimension of new embedding weights is not consistent with the old dimension'
-        model.gnn_model.features.weight = nn.Parameter(torch.FloatTensor(feature_mat), requires_grad=False)
+        model.gnn_model.features.weight = nn.Parameter(
+            torch.FloatTensor(feature_mat), requires_grad=False)
         model.gnn_model.features.num_embeddings = feature_mat.shape[0]
         model.gnn_model.features.embedding_dim = feature_mat.shape[1]
[…]
