+++ GraphSL/GNN/IVGD/correction.py
@@ -1,6 +1,7 @@
 import torch
 import torch.nn.functional as F
 
+
 class correction(torch.nn.Module):
     """
     Define an error correction module.
@@ -29,7 +30,8 @@ class correction(torch.nn.Module):
 
         - temp (torch.Tensor): Corrected prediction of the seed vector.
         """
-        # Concatenate the input tensor with its complement along the second dimension
+        # Concatenate the input tensor with its complement along the second
+        # dimension
         x = torch.cat((1 - x, x), dim=1)
         # Apply the first fully connected layer followed by ReLU activation
         temp = F.relu(self.fc1(x))
@@ -40,5 +42,13 @@ class correction(torch.nn.Module):
         # Add the input tensor to the output tensor
         temp = (temp + x)
         # Clip the values of the output tensor between 0 and 1
-        temp = torch.minimum(torch.maximum(torch.zeros(temp.shape).to(x.device), temp), torch.ones(temp.shape).to(x.device))
+        temp = torch.minimum(
+            torch.maximum(
+                torch.zeros(
+                    temp.shape).to(
+                    x.device),
+                temp),
+            torch.ones(
+                temp.shape).to(
+                x.device))
         return temp

[…]
