+++ requirements.txt
@@ -91,7 +91,7 @@ torch==2.2.2
 torch_geometric==2.5.2
 torchaudio==2.2.2
 torchvision==0.17.2
-tornado==6.4
+tornado==6.4.1
 tqdm==4.66.3
 traitlets==5.14.2
 triton==2.2.0
