+++ requirements.txt
@@ -78,7 +78,7 @@ pyzmq==25.1.2
 referencing==0.34.0
 requests==2.32.0
 rpds-py==0.18.0
-scikit-learn==1.4.1.post1
+scikit-learn==1.5.0
 scipy==1.12.0
 six==1.16.0
 soupsieve==2.5
