+++ requirements.txt
@@ -97,7 +97,7 @@ traitlets==5.14.2
 triton==2.2.0
 typing_extensions==4.10.0
 tzdata==2024.1
-urllib3==2.2.1
+urllib3==2.2.2
 wcwidth==0.2.13
 webencodings==0.5.1
 xyzservices==2023.10.1
