### 2024-07-04 xianggebenben/GraphSL#13 @mbeyss
[JOSS Review] Comments on paper.md
related to this [JOSS Review](https://github.com/openjournals/joss-reviews/issues/6796)

## From the JOSS checklist

- **Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?**

> The summary is hard to understand for a reader, who does not know the graph diffusion. I suggest do make the purpose a bit clearer, e.g. by putting the explanation what graph diffusion models are useful for ("simulating information spread") earlier and elaborating this more.

- **A statement of need: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?**

> Here the basic problem (graph source localization) is nicely introduced in a way that also non experts can understand . Figure 1 is rather useful for that.
The relation to others work becomes fairly clear, as you use other peoples software in a sort of unified framework.
For me, it is however not clear who the target audience is, is it developers for other algorithms for graph source localization, that want to compare their solution to others, or is it users of such algorithms, that want to find out the most suitable algorithm for their work? What I am missing for both cases is some steps how to do that, i.e. how to incorporate a new algorithm, or a new dataset. With that also the "problems the software is designed to solve" is a bit unclear, technically it solves the graph source localization problem, but it remains unclear what the exact use of GraphSL is, that other softwares in this field do not cover.

- **State of the field: Do the authors describe how this software compares to other commonly-used packages?**

> As it includes other packages, this is straight forward. However, as explained earlier, the usefulness of this (and the datasets) does not become clear

-  **Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?**

> The writing is generally of good quality. A few minor remarks:
 - line 27: "up-to-date state-of-the-art" seems redundant
 - in Methods and Benchmark Datasets the past tense is used, when describing other approaches, here present tense should be used.

- **References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?**

> This looks mostly fine. I would suggest to add reference to a recent review paper or so explaining graph source localization (techniques), to help the reader to dive deeper in to the field.
I feel like there are missing references for the data from table.

## More general

 - line 33:  "Independent Cascade" and "Linear Threshold" are not explained. references might also help
 - lines 33-36: the formulation here is not really clear. I would suggest to elaborate this further, This could go well with my earlier points
 - Figure 2:
   - the image has some sort of gradient to the left, background should either be white or transparent
   - the image description does not feel very useful. Suggestion: "The hierarchical structure of the GraphSL library. In total 6 algorithms are implemented, which can be devided into two categories: Prescribed Methods that rely on ... and GNN-based Methods which ..." From the image alone, along with the description the reader should be able to extract the most important information
- Table 1:
  - Caption has to be improved
  - The relevant information is lost in to much detail (two many numbers displayed, average Degree is redundant)
    - right align numbers and reduce number of significant digits (two, maybe three)
    - I would suggest to only have Nodes and Average Degree in there (average Degree needs to be fixed number of digits, as is, I would loose all digits after the decimal separator
- Line 55: There needs to be some sort of explanation of the datasets, what a seed deiffusion pair is and how to generate them.
