### 2024-07-04 xianggebenben/GraphSL#14 @mbeyss
[JOSS Review] Example Usage
related to this [JOSS Review](https://github.com/openjournals/joss-reviews/issues/6796)

While you have a very nice example that shows how to use all of your algorithms it could really need some comments and structure. I would suggest turning this into a jupyter notebook, where you can better explain and show what you are doing.
