### 2024-07-04 #6796 @mbeyss
Hi, I am looking for some guidance wrt to the review.
While there is a `test.py` file, that runs all the different algorithms, I do not see that it in some way checks if everything went well. What is the view on this? @rkdan deemed this acceptable, I am struggling a bit.
