### 2024-07-04 xianggebenben/GraphSL#15 @mbeyss
[Joss Review] Community guidelines
related to this [JOSS Review](https://github.com/openjournals/joss-reviews/issues/6796)

Please expand the `Contact` section of the README fot include

guidelines for third parties wishing to
1. Contribute to the software
2. Report issues or problems with the software
3. Seek support

3rd one seems covered and the other two are no brainers (pull request and issues in this repo). Just state it explicitly please.
