+++ paper.bib
@@ -165,3 +165,14 @@ year={2022}
   year={2012},
   publisher={Springer}
 }
+
+@article{jiang2016identifying,
+  title={Identifying propagation sources in networks: State-of-the-art and comparative studies},
+  author={Jiang, Jiaojiao and Wen, Sheng and Yu, Shui and Xiang, Yang and Zhou, Wanlei},
+  journal={IEEE Communications Surveys \& Tutorials},
+  volume={19},
+  number={1},
+  pages={465--481},
+  year={2016},
+  publisher={IEEE}
+}
