+++ paper.bib
@@ -176,3 +176,12 @@ year={2022}
   year={2016},
   publisher={IEEE}
 }
+
+@article{shakarian2015independent,
+  title={The independent cascade and linear threshold models},
+  author={Shakarian, Paulo and Bhatnagar, Abhivav and Aleali, Ashkan and Shaabani, Elham and Guo, Ruocheng and Shakarian, Paulo and Bhatnagar, Abhinav and Aleali, Ashkan and Shaabani, Elham and Guo, Ruocheng},
+  journal={Diffusion in social networks},
+  pages={35--48},
+  year={2015},
+  publisher={Springer}
+}
