+++ requirements.txt
@@ -6,7 +6,7 @@ backcall==0.2.0
 beautifulsoup4==4.12.3
 bleach==6.1.0
 bokeh==3.4.0
-certifi==2024.2.2
+certifi==2024.7.4
 charset-normalizer==3.3.2
 contourpy==1.2.0
 decorator==5.1.1
