### 2024-07-06 #6796 @xianggebenben
> @mbeyss - I would expect that tests have an expected result and the test checks that result is delivered. @xianggebenben can you add this? thanks!

Sure. Thanks.
