+++ README.md
@@ -35,14 +35,14 @@ GNN-based methods learn rules from graph data in an end-to-end manner by capturi
 
 |       Dataset      |  #Node |  #Edge | Average Degree | Seed-Diffusion Pairs |
 |:------------------:|:------:|:------:|:--------------:|:--------------------------:|
-|       Karate       |   34   |   78   |      4.588     |             No             |
-|      Dolphins      |   62   |   159  |      5.129     |             No             |
-|         Jazz       |   198  |  2,742 |     27.697     |             No             |
-| Network   Science  |  1,589 |  2,742 |      3.451     |             No             |
-|       Cora-ML      |  2,810 |  7,981 |      5.68      |             No             |
-|    Power   Grid    |  4,941 |  6,594 |      2.669     |             No             |
-|     Memetracker    |  7,884 | 47,911 |     12.154     |            Yes             |
-|        Digg        | 15,912 | 78,649 |      9.885     |            Yes             |
+|       Karate       |   34   |   78   |      5     |             No             |
+|      Dolphins      |   62   |   159  |      5     |             No             |
+|         Jazz       |   198  |  2,742 |     28     |             No             |
+| Network   Science  |  1,589 |  2,742 |      4     |             No             |
+|       Cora-ML      |  2,810 |  7,981 |      6     |             No             |
+|    Power   Grid    |  4,941 |  6,594 |      3     |             No             |
+|     Memetracker    |  7,884 | 47,911 |     12     |            Yes             |
+|        Digg        | 15,912 | 78,649 |      10     |            Yes             |
 
 
 Aside from methods, we also provide eight benchmark datasets to facilitate the research of the graph source localization problem. Memetracker and Digg provide Seed-Diffusion vector pairs. While others do not have such pairs, they can be generated by information diffusion simulations. All datasets are introduced as follows:
