+++ paper.md
@@ -58,12 +58,10 @@ GNN-based methods learn rules from graph data in an end-to-end manner by capturi
 | Network   Science [@newman2006finding] |  1,589 |  2,742 |
 |       Cora-ML [@mccallum2000automating]     |  2,810 |  7,981 |
 |    Power   Grid [@watts1998collective]    |  4,941 |  6,594 |
-|     Memetracker [@leskovec2009meme]    |  7,884 | 47,911 |
-|        Digg [@hogg2012social]       | 15,912 | 78,649 |
 
-Table: \label{tab:statistics}   Eight benchmark graph datasets: their numbers of nodes and edges.
+Table: \label{tab:statistics}   Six benchmark graph datasets: their numbers of nodes and edges.
 
-Aside from methods, we also release eight benchmark graph datasets to facilitate the research of graph source localization, whose statistics are shown in \autoref{tab:statistics}. While Memetracker and Digg provide information sources and diffusions, for other datasets, they can be generated by the function [diffusion_generation](https://graphsl.readthedocs.io/en/latest/GraphSL.html#GraphSL.utils.diffusion_generation).
+Aside from methods, we also release six benchmark graph datasets to facilitate the research of graph source localization, whose statistics are shown in \autoref{tab:statistics}. Information sources and diffusions can be generated by the function [diffusion_generation](https://graphsl.readthedocs.io/en/latest/GraphSL.html#GraphSL.utils.diffusion_generation).
 
 # Availability and Documentation
 
