+++ paper.md
@@ -38,7 +38,7 @@ Graph diffusion is a fundamental task in graph learning, which aims to predict f
 
 Some open-source tools have been developed to support the research of the graph source localization problem due to its importance. Two recent examples are cosasi [@McCabe2022joss] and RPaSDT [@frkaszczak2022rpasdt]. However, they missed comprehensive simulations of information diffusion, real-world benchmark datasets, and source localization approaches. To fill this gap, we propose a new library GraphSL: the first one to include both real-world benchmark datasets and recent source localization methods to our knowledge, which enables researchers and practitioners to easily evaluate novel techniques against appropriate baselines. These methods do not require prior knowledge (e.g. single source or multiple sources), and can handle graph source localization based on various diffusion simulation models such as Independent Cascade (IC) and Linear Threshold (LT) [@shakarian2015independent]. Our GraphSL library is standardized: for instance, tests of all source inference methods return a Metric object, which provides five performance metrics (accuracy, precision, recall, F-score, and area under ROC curve) for performance evaluation.
 
-Our proposed GraphSL library targets at both developers and practical users: they are free to add algorithms and datasets for personal needs by following the guidelines of in the "Contributions to GraphSL" section of [README.md](https://github.com/xianggebenben/GraphSL/blob/main/README.md).
+Our proposed GraphSL library targets at both developers and practical users: they are free to add algorithms and datasets for personal needs by following the guidelines of in the "Contact" section of [README.md](https://github.com/xianggebenben/GraphSL/blob/main/README.md).
 
 # Methods and Benchmark Datasets
 
