+++ README.md
@@ -33,16 +33,14 @@ GNN-based methods learn rules from graph data in an end-to-end manner by capturi
 
 ## Benchmark Datasets
 
-|       Dataset      |  #Node |  #Edge | Average Degree | Seed-Diffusion Pairs |
-|:------------------:|:------:|:------:|:--------------:|:--------------------------:|
-|       Karate       |   34   |   78   |      5     |             No             |
-|      Dolphins      |   62   |   159  |      5     |             No             |
-|         Jazz       |   198  |  2,742 |     28     |             No             |
-| Network   Science  |  1,589 |  2,742 |      4     |             No             |
-|       Cora-ML      |  2,810 |  7,981 |      6     |             No             |
-|    Power   Grid    |  4,941 |  6,594 |      3     |             No             |
-|     Memetracker    |  7,884 | 47,911 |     12     |            Yes             |
-|        Digg        | 15,912 | 78,649 |      10     |            Yes             |
+|       Dataset      |  #Node |  #Edge |
+|:------------------:|:------:|:------:|
+|       Karate       |   34   |   78   |
+|      Dolphins      |   62   |   159  |
+|         Jazz       |   198  |  2,742 |
+| Network   Science  |  1,589 |  2,742 |
+|       Cora-ML      |  2,810 |  7,981 |
+|    Power   Grid    |  4,941 |  6,594 |
 
 
 Aside from methods, we also provide eight benchmark datasets to facilitate the research of the graph source localization problem. Memetracker and Digg provide Seed-Diffusion vector pairs. While others do not have such pairs, they can be generated by information diffusion simulations. All datasets are introduced as follows:
@@ -150,5 +148,5 @@ If you use this package in your research, please consider citing our work as fol
 }
 ```
 # Contact
-
-Feel free to Email Junxiang Wang (junxiang.wang@alumni.emory.edu) if you have any questions.
+We welcome your contributions! If you’d like to contribute your datasets or algorithms, please submit a pull request consisting of an atomic commit and a brief message describing your contribution.
+Feel free to Email me (junxiang.wang@alumni.emory.edu) if you have any questions. Bug reports and feedback can be directed to the [Github issues page](https://github.com/xianggebenben/GraphSL/issues).
