+++ README.md
@@ -43,7 +43,7 @@ GNN-based methods learn rules from graph data in an end-to-end manner by capturi
 |    Power   Grid    |  4,941 |  6,594 |
 
 
-Aside from methods, we also provide eight benchmark datasets to facilitate the research of the graph source localization problem. Memetracker and Digg provide Seed-Diffusion vector pairs. While others do not have such pairs, they can be generated by information diffusion simulations. All datasets are introduced as follows:
+Aside from methods, we also provide six benchmark datasets to facilitate the research of the graph source localization problem. All datasets are introduced as follows:
 
 1. Karate: Karate depicts the social ties among members of a university karate club.
 
