+++ test.py
@@ -8,9 +8,9 @@ curr_dir = os.getcwd()
 # 'karate', 'dolphins', 'jazz', 'netscience', 'cora_ml', 'power_grid'
 data_name = 'karate'
 graph = load_dataset(data_name, data_dir=curr_dir)
-dataset = diffusion_generation(graph=graph, infect_prob=0.3, diff_type='IC', sim_num=100, seed_ratio=0.2)
+dataset = diffusion_generation(graph=graph, infect_prob=0.3, diff_type='IC', sim_num=10, seed_ratio=0.2)
 adj, train_dataset, test_dataset = split_dataset(dataset)
-lpsi = LPSI()
+"""lpsi = LPSI()
 alpha, thres, auc, f1, pred = lpsi.train(adj, train_dataset)
 print("LPSI:")
 print(f"train auc: {auc:.3f}, train f1: {f1:.3f}")
@@ -22,30 +22,30 @@ print("NetSleuth:")
 print(f"train auc: {auc:.3f}, train f1: {f1:.3f}")
 metric = netSleuth.test(adj, test_dataset, k)
 print(f"test acc: {metric.acc:.3f}, test pr: {metric.pr:.3f}, test re: {metric.re:.3f}, test f1: {metric.f1:.3f}, test auc: {metric.auc:.3f}")
+print("OJC:")
 ojc = OJC()
 Y, auc, f1 = ojc.train(adj, train_dataset)
-print("OJC:")
 print(f"train auc: {auc:.3f}, train f1: {f1:.3f}")
 metric = ojc.test(adj, test_dataset, Y)
 print(f"test acc: {metric.acc:.3f}, test pr: {metric.pr:.3f}, test re: {metric.re:.3f}, test f1: {metric.f1:.3f}, test auc: {metric.auc:.3f}")
+print("GCNSI:")
 gcnsi = GCNSI()
 gcnsi_model, thres, auc, f1, pred = gcnsi.train(adj, train_dataset)
-print("GCNSI:")
 print(f"train auc: {auc:.3f}, train f1: {f1:.3f}")
 metric = gcnsi.test(adj, test_dataset, gcnsi_model, thres)
-print(f"test acc: {metric.acc:.3f}, test pr: {metric.pr:.3f}, test re: {metric.re:.3f}, test f1: {metric.f1:.3f}, test auc: {metric.auc:.3f}")
+print(f"test acc: {metric.acc:.3f}, test pr: {metric.pr:.3f}, test re: {metric.re:.3f}, test f1: {metric.f1:.3f}, test auc: {metric.auc:.3f}")"""
+print("IVGD:")
 ivgd = IVGD()
 diffusion_model = ivgd.train_diffusion(adj, train_dataset)
 ivgd_model, thres, auc, f1, pred = ivgd.train(
     adj, train_dataset, diffusion_model)
-print("IVGD:")
 print(f"train auc: {auc:.3f}, train f1: {f1:.3f}")
 metric = ivgd.test(test_dataset, diffusion_model, ivgd_model, thres)
 print(f"test acc: {metric.acc:.3f}, test pr: {metric.pr:.3f}, test re: {metric.re:.3f}, test f1: {metric.f1:.3f}, test auc: {metric.auc:.3f}")
+print("SLVAE:")
 slave = SLVAE()
 slvae_model, seed_vae_train, thres, auc, f1, pred = slave.train(
     adj, train_dataset)
-print("SLVAE:")
 print(f"train auc: {auc:.3f}, train f1: {f1:.3f}")
 metric = slave.infer(test_dataset, slvae_model, seed_vae_train, thres)
[…]

[…]
