+++ GraphSL/GNN/IVGD/main.py
@@ -32,30 +32,21 @@ class IVGD_model(torch.nn.Module):
         - rho (float): Value of rho parameter.
         """
         super(IVGD_model, self).__init__()
-        self.number_layer = 5
         self.alpha1 = alpha
         self.alpha2 = alpha
         self.alpha3 = alpha
-        self.alpha4 = alpha
-        self.alpha5 = alpha
 
         self.tau1 = tau
         self.tau2 = tau
         self.tau3 = tau
-        self.tau4 = tau
-        self.tau5 = tau
 
         self.net1 = correction()
         self.net2 = correction()
         self.net3 = correction()
-        self.net4 = correction()
-        self.net5 = correction()
 
         self.rho1 = rho
         self.rho2 = rho
         self.rho3 = rho
-        self.rho4 = rho
-        self.rho5 = rho
 
     def forward(self, x, label, lamda):
         """
@@ -76,8 +67,6 @@ class IVGD_model(torch.nn.Module):
         self.net1.to(x.device)
         self.net2.to(x.device)
         self.net3.to(x.device)
-        self.net4.to(x.device)
-        self.net5.to(x.device)
         sum = torch.sum(label)
         label = torch.cat((1 - label, label), dim=1)
         x = torch.cat((1 - x, x), dim=1)
@@ -95,33 +84,7 @@ class IVGD_model(torch.nn.Module):
         x = (self.tau3 * self.net3(prob) - label * torch.softmax(x, dim=1) / label.shape[0] - lamda
              - self.rho3 * (torch.sum(x) - sum) + self.alpha3 * x) / (
             self.tau3 + self.alpha3)
-        prob = x[:, 1].unsqueeze(-1)
-        lamda = lamda + self.rho3 * (torch.sum(prob) - sum)
-        x = (self.tau4 * self.net4(prob) - label * torch.softmax(x, dim=1) / label.shape[0] - lamda
-             - self.rho4 * (torch.sum(x) - sum) + self.alpha4 * x) / (
-            self.tau4 + self.alpha4)
-        prob = x[:, 1].unsqueeze(-1)
-        lamda = lamda + self.rho4 * (torch.sum(prob) - sum)
-        x = (self.tau5 * self.net5(prob) - label * torch.softmax(x, dim=1) / label.shape[0] - lamda
-             - self.rho5 * (torch.sum(x) - sum) + self.alpha5 * x) / (
-            self.tau5 + self.alpha5)
         return x
-
-    def correction(self, pred):
-        """
-        Corrects predictions based on input.
-
-        Args:
-
-        - pred (torch.Tensor): Input tensor of predictions.
-
-        Returns:
-
-        - torch.Tensor: Corrected predictions.
-        """
-        temp = pred[:, 0].unsqueeze(-1)
-        return (self.net1(temp) + self.net2(temp) + self.net3(temp) +
-                self.net4(temp) + self.net5(temp)) / self.number_layer
 
 
 class IVGD:
@@ -311,7 +274,7 @@ class IVGD:
         alpha = 1
         tau = 1
         rho = 1e-3
-        lamda = 0
+        lamda = 1e-3
         torch.manual_seed(random_seed)
         ivgd = IVGD_model(alpha=alpha, tau=tau, rho=rho)
         optimizer = optim.Adam(
@@ -320,29 +283,63 @@ class IVGD:
             weight_decay=weight_decay)
         ivgd.train()
         train_num = len(train_dataset)
-        for i, influ_mat in enumerate(train_dataset):
-            seed_vec = influ_mat[:, 0].to(device)
+
+        seed_preds_list=list()
+
+        for influ_mat in train_dataset:
+
             influ_vec = influ_mat[:, -1].to(device)
+
[…]
