+++ test.py
@@ -8,9 +8,9 @@ curr_dir = os.getcwd()
 # 'karate', 'dolphins', 'jazz', 'netscience', 'cora_ml', 'power_grid'
 data_name = 'karate'
 graph = load_dataset(data_name, data_dir=curr_dir)
-dataset = diffusion_generation(graph=graph, infect_prob=0.3, diff_type='IC', sim_num=10, seed_ratio=0.2)
+dataset = diffusion_generation(graph=graph, infect_prob=0.3, diff_type='IC', sim_num=100, seed_ratio=0.2)
 adj, train_dataset, test_dataset = split_dataset(dataset)
-"""lpsi = LPSI()
+lpsi = LPSI()
 alpha, thres, auc, f1, pred = lpsi.train(adj, train_dataset)
 print("LPSI:")
 print(f"train auc: {auc:.3f}, train f1: {f1:.3f}")
@@ -33,7 +33,7 @@ gcnsi = GCNSI()
 gcnsi_model, thres, auc, f1, pred = gcnsi.train(adj, train_dataset)
 print(f"train auc: {auc:.3f}, train f1: {f1:.3f}")
 metric = gcnsi.test(adj, test_dataset, gcnsi_model, thres)
-print(f"test acc: {metric.acc:.3f}, test pr: {metric.pr:.3f}, test re: {metric.re:.3f}, test f1: {metric.f1:.3f}, test auc: {metric.auc:.3f}")"""
+print(f"test acc: {metric.acc:.3f}, test pr: {metric.pr:.3f}, test re: {metric.re:.3f}, test f1: {metric.f1:.3f}, test auc: {metric.auc:.3f}")
 print("IVGD:")
 ivgd = IVGD()
 diffusion_model = ivgd.train_diffusion(adj, train_dataset)
