+++ docs/index.rst
@@ -25,17 +25,17 @@ First, install GraphSL using pip:
 
     pip install GraphSL
 
-Or, clone the repo from https://github.com/xianggebenben/GraphSL, and install requirements:
+Or, clone the [repo](https://github.com/xianggebenben/GraphSL), and install requirements:
 
 .. code-block:: bash
 
     pip install -r requirements.txt
 
-Second, download the data folder of the repo from https://github.com/xianggebenben/GraphSL, which contains six datasets.
+Second, download the data folder of the [repo](https://github.com/xianggebenben/GraphSL), which contains six datasets.
 
 
 Usage
------
+------------
 
 Now, you can import and use GraphSL in your Python code. Please make sure your Python code and the data folder are in the same directory:
 
@@ -136,7 +136,7 @@ Now, you can import and use GraphSL in your Python code. Please make sure your P
   
 
 
-That's it! You're ready to start using GraphSL.
+That's it! You're ready to start using GraphSL. You check results on the Jupyter notebook tutorial.ipynb from the [repo](https://github.com/xianggebenben/GraphSL).
 
 If you use this package in your research, please consider citing our work as follows:
 
@@ -149,6 +149,16 @@ If you use this package in your research, please consider citing our work as fol
   journal = {preprint, 	arXiv:2405.03724}
    }
 
+Contact
+------------
+
+We welcome your contributions! If you’d like to contribute your datasets or algorithms, please submit a pull request consisting of an atomic commit and a brief message describing your contribution.
+
+For a new dataset, please upload it to the [data](https://github.com/xianggebenben/GraphSL/tree/main/data) folder. The file should be a dictionary object saved by [pickle](https://docs.python.org/3/library/pickle.html). It contains a key "adj_mat" with the value of a garph adjacency matrix (sprase numpy array with the [CSR](https://docs.scipy.org/doc/scipy/reference/generated/scipy.sparse.csr_matrix.html) format).
+
+For a new algorithm, please determine it whether it belongs to presribed methods or GNN-based methods: if it belongs to the prescribed methods,  add your algorithm as a new class in the [GraphSL/Prescribed.py](https://github.com/xianggebenben/GraphSL/blob/main/GraphSL/Prescribed.py). Otherwises, please upload it as a folder under the [GraphSL/GNN](https://github.com/xianggebenben/GraphSL/tree/main/GraphSL/GNN) folder. Typically, the algorithm should include a "train" function and a "test" function, and the "test" function should return a [Metric](https://github.com/xianggebenben/GraphSL/blob/main/GraphSL/Evaluation.py) object.
+
+Feel free to Email me (junxiang.wang@alumni.emory.edu) if you have any questions. Bug reports and feedback can be directed to the [Github issues page](https://github.com/xianggebenben/GraphSL/issues).
 
 Indices and tables
 ==================
