+++ build/lib/GraphSL/GNN/IVGD/main.py
@@ -7,12 +7,11 @@ from GraphSL.GNN.IVGD.i_deepis import i_DeepIS, DiffusionPropagate
 from GraphSL.GNN.IVGD.training import FeatureCons, get_idx_new_seeds
 from GraphSL.GNN.IVGD.model.MLP import MLPTransform
 from GraphSL.GNN.IVGD.training import train_model
-from sklearn.metrics import roc_auc_score,f1_score,accuracy_score,precision_score,recall_score
+from sklearn.metrics import roc_auc_score, f1_score, accuracy_score, precision_score, recall_score
 from GraphSL.Evaluation import Metric
 import torch.optim as optim
 import warnings
 warnings.filterwarnings("ignore")
-
 
 
 class IVGD_model(torch.nn.Module):
@@ -33,30 +32,21 @@ class IVGD_model(torch.nn.Module):
         - rho (float): Value of rho parameter.
         """
         super(IVGD_model, self).__init__()
-        self.number_layer = 5
         self.alpha1 = alpha
         self.alpha2 = alpha
         self.alpha3 = alpha
-        self.alpha4 = alpha
-        self.alpha5 = alpha
 
         self.tau1 = tau
         self.tau2 = tau
         self.tau3 = tau
-        self.tau4 = tau
-        self.tau5 = tau
 
         self.net1 = correction()
         self.net2 = correction()
         self.net3 = correction()
-        self.net4 = correction()
-        self.net5 = correction()
 
         self.rho1 = rho
         self.rho2 = rho
         self.rho3 = rho
-        self.rho4 = rho
-        self.rho5 = rho
 
     def forward(self, x, label, lamda):
         """
@@ -77,52 +67,24 @@ class IVGD_model(torch.nn.Module):
         self.net1.to(x.device)
         self.net2.to(x.device)
         self.net3.to(x.device)
-        self.net4.to(x.device)
-        self.net5.to(x.device)
         sum = torch.sum(label)
         label = torch.cat((1 - label, label), dim=1)
         x = torch.cat((1 - x, x), dim=1)
         prob = x[:, 1].unsqueeze(-1)
         x = (self.tau1 * self.net1(prob) - label * torch.softmax(x, dim=1) / label.shape[0] - lamda
              - self.rho1 * (torch.sum(x) - sum) + self.alpha1 * x) / (
-                    self.tau1 + self.alpha1)
+            self.tau1 + self.alpha1)
         prob = x[:, 1].unsqueeze(-1)
         lamda = lamda + self.rho1 * (torch.sum(prob) - sum)
         x = (self.tau2 * self.net2(prob) - label * torch.softmax(x, dim=1) / label.shape[0] - lamda
              - self.rho2 * (torch.sum(x) - sum) + self.alpha2 * x) / (
-                    self.tau2 + self.alpha2)
+            self.tau2 + self.alpha2)
         prob = x[:, 1].unsqueeze(-1)
         lamda = lamda + self.rho2 * (torch.sum(prob) - sum)
         x = (self.tau3 * self.net3(prob) - label * torch.softmax(x, dim=1) / label.shape[0] - lamda
              - self.rho3 * (torch.sum(x) - sum) + self.alpha3 * x) / (
-                    self.tau3 + self.alpha3)
-        prob = x[:, 1].unsqueeze(-1)
-        lamda = lamda + self.rho3 * (torch.sum(prob) - sum)
-        x = (self.tau4 * self.net4(prob) - label * torch.softmax(x, dim=1) / label.shape[0] - lamda
-             - self.rho4 * (torch.sum(x) - sum) + self.alpha4 * x) / (
-                    self.tau4 + self.alpha4)
[…]
