+++ build/lib/GraphSL/GNN/SLVAE/main.py
@@ -2,10 +2,12 @@ import torch.nn as nn
 import torch
 import torch.nn.functional as F
 import numpy as np
-from GraphSL.GNN.SLVAE.model import VAE,GNN,DiffusionPropagate
+from GraphSL.GNN.SLVAE.model import VAE, GNN, DiffusionPropagate
 from torch.optim import Adam
-from sklearn.metrics import roc_auc_score,f1_score,accuracy_score,precision_score,recall_score
+from sklearn.metrics import roc_auc_score, f1_score, accuracy_score, precision_score, recall_score
 from GraphSL.Evaluation import Metric
+
+
 class SLVAE_model(nn.Module):
     """
     Source Localization Variational Autoencoder (SLVAE) model combining VAE, GNN, and propagation modules.
@@ -35,7 +37,10 @@ class SLVAE_model(nn.Module):
         self.vae = vae
         self.gnn = gnn
         self.propagate = propagate
-        self.reg_params = list(filter(lambda x: x.requires_grad, self.gnn.parameters()))
+        self.reg_params = list(
+            filter(
+                lambda x: x.requires_grad,
+                self.gnn.parameters()))
 
     def forward(self, seed_vec, train_mode):
         """
@@ -57,7 +62,8 @@ class SLVAE_model(nn.Module):
 
         - predictions (torch.Tensor): Predictions made by the SLVAE model.
         """
-        # Pass seed_vec through VAE to obtain reconstructed seed vector, mean, and log variance
+        # Pass seed_vec through VAE to obtain reconstructed seed vector, mean,
+        # and log variance
         seed_hat, mean, log_var = self.vae(seed_vec)
 
         if train_mode:
@@ -124,6 +130,7 @@ class SLVAE_model(nn.Module):
 
         - total_loss (torch.Tensor): Total loss tensor.
         """
+        epsilon =1e-8
         device = y_true.device
         BN = nn.BatchNorm1d(1, affine=False).to(device)
         y_hat = y_hat.to(device)
@@ -132,7 +139,8 @@ class SLVAE_model(nn.Module):
         for pred in train_pred:
             log_lh = torch.zeros(1).to(device)
             for i, x_i in enumerate(x_hat[0]):
-                temp = x_i * torch.log(pred[i]) + (1 - x_i) * torch.log(1 - pred[i]).to(torch.double)
+                temp = x_i * \
+                    torch.log(pred[i]+epsilon) + (1 - x_i) * torch.log(1 - pred[i]+epsilon).to(torch.double)
                 temp = temp.to(device)
                 log_lh += temp
             log_pmf.append(log_lh)
@@ -148,6 +156,7 @@ class SLVAE_model(nn.Module):
 
         return total_loss
 
+
 class SLVAE:
     """
     Implement the Source Localization Variational Autoencoder (SLVAE) model.
@@ -160,7 +169,21 @@ class SLVAE:
         Initialize the SLVAE model.
         """
 
-    def train(self, adj, train_dataset, thres_list=[0.1, 0.3, 0.5, 0.7, 0.9], lr = 1e-3, weight_decay = 1e-4, num_epoch = 50,print_epoch =10):
+    def train(
+            self,
+            adj,
+            train_dataset,
+            thres_list=[
+                0.1,
+                0.3,
+                0.5,
+                0.7,
+                0.9],
+            lr=1e-4,
+            weight_decay=1e-4,
+            num_epoch=100,
+            print_epoch=10,
+            random_seed=0):
         """
         Train the SLVAE model.
 
@@ -179,6 +202,8 @@ class SLVAE:
         - num_epoch (int): Number of training epochs.
 
         - print_epoch (int): Number of epochs every time to print loss.
+
+        - random_seed (int): Random seed.
 
[…]
