+++ build/lib/GraphSL/utils.py
@@ -13,7 +13,7 @@ def load_dataset(dataset, data_dir):
 
     Args:
 
-    - dataset (str): The name of the dataset file, 'karate', 'dolphins', 'jazz', 'netscience', 'cora_ml', 'power_grid','meme8000', 'digg16000'.
+    - dataset (str): The name of the dataset file, 'karate', 'dolphins', 'jazz', 'netscience', 'cora_ml', 'power_grid'.
 
     - data_dir (str): The directory where the dataset files are stored.
 
