### 2024-07-10 xianggebenben/GraphSL#14 @xianggebenben
Thank you for the suggestion. We have added a jupyter notebook file [tutorial.ipynb](https://github.com/xianggebenben/GraphSL/blob/main/tutorial.ipynb) to explain the implementation.
