+++ docs/index.rst
@@ -25,13 +25,13 @@ First, install GraphSL using pip:
 
     pip install GraphSL
 
-Or, clone the repo(https://github.com/xianggebenben/GraphSL), and install requirements:
+Or, clone the repo (https://github.com/xianggebenben/GraphSL), and install requirements:
 
 .. code-block:: bash
 
     pip install -r requirements.txt
 
-Second, download the data folder of the repo(https://github.com/xianggebenben/GraphSL), which contains six datasets.
+Second, download the data folder of the repo (https://github.com/xianggebenben/GraphSL), which contains six datasets.
 
 
 Usage
