### 2024-07-10 #6796 @xianggebenben
> Also, judging Contribution and authorship is difficult as the main software contribution is made in a single commit by @xianggebenben. After that there are countless commits with mostly technical contents by @xianggebenben, who certainly earned their spot on the paper, but it is unclear to what extend they or others (e.g. Liang Zhaos, or others that are not mentioned) have contributed to the software.

Thank you for the comments. We have revised the paper based on your suggestions, and we also replied to your comments to address concerns.  Please let me know if you have additional comments.

For contribution and authorship, I wrote the whole GraphSL library, and drafted the paper and documentations, while my co-author Liang Zhao provided the datasets. Please let me know if I need to add them to the paper or readme file.
