### 2024-07-10 #6796 @xianggebenben
> @mbeyss - I would expect that tests have an expected result and the test checks that result is delivered. @xianggebenben can you add this? thanks!

I have added a jupyter notebook file [tutorial.ipynb](https://github.com/xianggebenben/GraphSL/blob/main/tutorial.ipynb) to print the results, which can be utilized to check the expected outputs. Our library can set seeds and make implementations reproducible. Please let me know if you have any suggestions.
