+++ GraphSL/GNN/GCNSI/main.py
@@ -132,6 +132,7 @@ class GCNSI:
             d4 = d4[:, np.newaxis]
             lpsi_input[i, :, :] = np.concatenate((d1, d2, d3, d4), axis=1)
         # Training loop
+        print("train GCNSI:")
         for epoch in range(num_epoch):
             optimizer.zero_grad()
             total_loss = 0
@@ -146,7 +147,7 @@ class GCNSI:
                 optimizer.step()
             average_loss = total_loss / train_num
             if epoch % print_epoch == 0:
-                print(f"epoch = {epoch}, loss = {average_loss:.3f}")
+                print(f"Epoch ({epoch}/{num_epoch}), loss = {average_loss:.3f}")
         train_auc = 0
         # Compute AUC score on training data
         for i, influ_mat in enumerate(train_dataset):
