### 2024-07-15 #6796 @xianggebenben
> Hi I have reviewed the paper @majensen. I'm attaching my comments below: Im revising this a bit late so I suppose all the edits have been made. Here are my comments @xianggebenben.
>
> * I appreciate the Quickstart to test the repo
> * Not an immediate requirement, but after I pip installed into an environment, the datasets have to be separate folder (which Is mentioned, kudos) but it would be nice if it had an interface like NLTK where we can install datasets internally from the module
> * I don’t see any particular issues with the contributions of the authors
> * Edits to the paper:
>
> Line 14 “its inverse problem, graph source localization, though rarely explored, stands as an extremely important topic" * Consider rephrasing for clarity: "its inverse problem, graph source localization, although rarely explored, is an extremely important topic".
>
> Line 19 “spans a broad spectrum of promising research and real-world applications such as rumor detection (Evanega et al., 2020), tracking of sources for computer viruses, (Kephart & White, 1993), and failures detection in smart grids (Amin & Schewe, 2007)." * Remove the comma after "computer viruses". * Change "failures detection" to "failure detection".
>
> Line 31 “methods do not require prior knowledge (e.g. single source or multiple sources)," * Consider changing "prior knowledge" to "prior assumptions about the source (e.g., single source or multiple sources),".
>
> Line 41 “Prescribed methods rely on hand-crafted rules and heuristics. For instance, LPSI propagated41 infection in networks and labels local peaks as source nodes" * Consider rephrasing for clarity: "Prescribed methods rely on hand-crafted rules and heuristics. For instance, LPSI propagates infections in networks and labels local peaks as source nodes".

Thank you for the comments @range-et. Our reply is shown as follows:

Q1.  Not an immediate requirement, but after I pip installed into an environment, the datasets have to be separate folder (which Is mentioned, kudos) but it would be nice if it had an interface like NLTK where we can install datasets internally from the module.

A1. Thank you for the suggestion. We have added a new function download_dataset to achieve this goal. It is shown in the [tutorial](https://github.com/xianggebenben/GraphSL/blob/main/tutorial.ipynb).

Q2. Edits to the paper.
A2. We have revised the paper based on your comments. Thank you.
Please let me know if you have any additional comments. Have a good day.
