### 2024-07-16 xianggebenben/GraphSL#14 @xianggebenben
> That notebook is a good first step. What I hoped for was some added comments. Notebooks offer great ways of structuring and commenting code

Thank you for the comment. The jupyter notebook file [tutorial.ipynb](https://github.com/xianggebenben/GraphSL/blob/main/tutorial.ipynb) is added with detailed comments of explanations. Please let me know if you have any questions.
