### 2024-07-17 xianggebenben/GraphSL#14 @mbeyss
I think the features of a notebook can be better employed. I created a pull request, to show what I mean. Have a look at #17
