### 2024-07-17 #6796 @mbeyss
> > @mbeyss - I would expect that tests have an expected result and the test checks that result is delivered. @xianggebenben can you add this? thanks!
>
> I have added a jupyter notebook file [tutorial.ipynb](https://github.com/xianggebenben/GraphSL/blob/main/tutorial.ipynb) to print the results, which can be utilized to check the expected outputs. Our library can set seeds and make implementations reproducible. Please let me know if you have any suggestions.

I am still struggling a bit with that. The expected result is that the notebook runs without errors, which is a not to helpful.
But I also do not want to nitpick. Perhaps a comment / explanation on the metrics returned from the test would be useful: what do accuracy, precision, recall, F1-Score and area under the ROC curve mean? What are good values, how can I judge if the test was successful?

Apart from that, I strongly encourage the authors to implement automated testing, the CI-pipeline is already there.
