### 2024-07-17 xianggebenben/GraphSL#14 @xianggebenben
> I think the features of a notebook can be better employed. I created a pull request, to show what I mean. Have a look at #17

Thanks very much for your help. I have merged the pull request.
