+++ requirements.txt
@@ -104,3 +104,4 @@ xyzservices==2023.10.1
 yarg==0.1.9
 yarl==1.9.4
 sphinx-rtd-theme==2.0.0
+pytest==8.2.2
