+++ test.py
@@ -2,7 +2,7 @@ from GraphSL.GNN.SLVAE.main import SLVAE
 from GraphSL.GNN.IVGD.main import IVGD
 from GraphSL.GNN.GCNSI.main import GCNSI
 from GraphSL.Prescribed import LPSI, NetSleuth, OJC
-from GraphSL.utils import load_dataset, diffusion_generation, split_dataset,download_dataset
+from GraphSL.utils import load_dataset, diffusion_generation, split_dataset,download_dataset,visualize_source_prediction
 import os
 curr_dir = os.getcwd()
 # download datasets
@@ -59,6 +59,11 @@ gcnsi = GCNSI()
 gcnsi_model, thres, auc, f1, pred = gcnsi.train(adj, train_dataset)
 print(f"train auc: {auc:.3f}, train f1: {f1:.3f}")
 
+# visualize training predictions
+pred = (pred >= thres)
+visualize_source_prediction(adj,pred[:,0],train_dataset[0][:,0].numpy(),save_dir=curr_dir,save_name="GCNSI_source_prediction")
+
+
 # test GCNSI
 metric = gcnsi.test(adj, test_dataset, gcnsi_model, thres)
 print(f"test acc: {metric.acc:.3f}, test pr: {metric.pr:.3f}, test re: {metric.re:.3f}, test f1: {metric.f1:.3f}, test auc: {metric.auc:.3f}")
@@ -75,6 +80,10 @@ ivgd_model, thres, auc, f1, pred = ivgd.train(
     adj, train_dataset, diffusion_model)
 print(f"train auc: {auc:.3f}, train f1: {f1:.3f}")
 
+# visualize training predictions
+pred = (pred >= thres)
+visualize_source_prediction(adj,pred[:,0],train_dataset[0][:,0].numpy(),save_dir=curr_dir,save_name="IVGD_source_prediction")
+
 # test IVGD
 metric = ivgd.test(test_dataset, diffusion_model, ivgd_model, thres)
 print(f"test acc: {metric.acc:.3f}, test pr: {metric.pr:.3f}, test re: {metric.re:.3f}, test f1: {metric.f1:.3f}, test auc: {metric.auc:.3f}")
@@ -87,6 +96,10 @@ slave = SLVAE()
 slvae_model, seed_vae_train, thres, auc, f1, pred = slave.train(
     adj, train_dataset)
 print(f"train auc: {auc:.3f}, train f1: {f1:.3f}")
+
+# visualize training predictions
+pred = (pred >= thres)
+visualize_source_prediction(adj,pred[:,0],train_dataset[0][:,0].numpy(),save_dir=curr_dir,save_name="SLVAE_source_prediction")
 
 # test SLVAE
 metric = slave.infer(test_dataset, slvae_model, seed_vae_train, thres)
