### 2024-07-23 97f320b
use num_thres instead of thres_list
