+++ README.md
@@ -213,3 +213,5 @@ Version 0.11 removes the memetracker and the digg datasets, improves the IVGD me
 Version 0.12 adds the datasets downloader.
 
 Version 0.13 adds the visualization of source predictions.
+
+Version 0.14 uses the num_thres (i.e. number of thresholds to try) instead of specificing thres_list (i.e. threshold list) for LPSI, GCNSI, IVGD and SLVAE. Moreover, GCNSI, IVGD and SLVAE are improved to run on CUDA if applicable.
