### 2024-07-25 #6796 @majensen
Hi @xianggebenben - can you please add the DOI references [as suggested by editorialbot](https://github.com/openjournals/joss-reviews/issues/6796#issuecomment-2250341483)?

Regarding the text, I have to very minor suggestions:
l. 27 - “to support the research” -> “to support research”
l. 40 - “Our proposed GraphSL” -> “Our GraphSL”

Thanks
