### 2024-07-25 befb4f0
Update paper.md
