+++ paper.bib
@@ -13,7 +13,8 @@
   booktitle={Proceedings 1993 IEEE Computer Society Symposium on Research in Security and Privacy},
   pages={2--15},
   year={1993},
-  organization={IEEE}
+  organization={IEEE},
+doi={10.1109/risp.1993.287647}
 }
 @article{amin2007preventing,
   title={Preventing blackouts},
@@ -23,14 +24,16 @@
   number={5},
   pages={60--67},
   year={2007},
-  publisher={JSTOR}
+  publisher={JSTOR},
+  doi={10.1049/pe:20030305}
 }
 
 @inproceedings{IVGD_www22,
 title={An Invertible Graph Diffusion Neural Network for Source Localization},
 author={Wang, Junxiang and Jiang, Junji and Zhao, Liang},
 booktitle={Proceedings of the 31th International World Wide Web Conference (WWW 2022)},
-year={2022}
+year={2022},
+doi={10.1145/3485447.3512155}
 
 }
 @inproceedings{wang2017multiple,
@@ -39,7 +42,8 @@ year={2022}
   booktitle={Proceedings of the AAAI Conference on Artificial Intelligence},
   volume={31},
   number={1},
-  year={2017}
+  year={2017},
+  doi={10.1609/aaai.v31i1.10477}
 }
 @inproceedings{zhu2017catch,
   title={Catch’em all: Locating multiple diffusion sources in networks with partial observations},
@@ -47,7 +51,8 @@ year={2022}
   booktitle={Proceedings of the AAAI Conference on Artificial Intelligence},
   volume={31},
   number={1},
-  year={2017}
+  year={2017},
+  doi={10.1609/aaai.v31i1.10746}
 }
 @inproceedings{prakash2012spotting,
   title={Spotting culprits in epidemics: How many and which ones?},
@@ -55,21 +60,24 @@ year={2022}
   booktitle={2012 IEEE 12th international conference on data mining},
   pages={11--20},
   year={2012},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/icdm.2012.136}
 }
 @inproceedings{dong2019multiple,
   title={Multiple rumor source detection with graph convolutional networks},
   author={Dong, Ming and Zheng, Bolong and Quoc Viet Hung, Nguyen and Su, Han and Li, Guohui},
   booktitle={Proceedings of the 28th ACM international conference on information and knowledge management},
   pages={569--578},
-  year={2019}
+  year={2019},
+  doi = {10.1145/3357384.3357994}
 }
 @inproceedings{ling2022source,
   title={Source localization of graph diffusion via variational autoencoders for graph inverse problems},
   author={Ling, Chen and Jiang, Junji and Wang, Junxiang and Liang, Zhao},
   booktitle={Proceedings of the 28th ACM SIGKDD conference on knowledge discovery and data mining},
   pages={1010--1020},
-  year={2022}
+  year={2022},
+  doi={10.1145/3534678.3539288}
 }
 
 @article{McCabe2022joss,
@@ -91,7 +99,8 @@ year={2022}
   volume={17},
   pages={100988},
   year={2022},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.softx.2022.100988}
 }
 
 @article{lusseau2003bottlenose,
@@ -101,7 +110,8 @@ year={2022}
   volume={54},
   pages={396--405},
   year={2003},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s00265-003-0651-y}
 }
 
 @article{gleiser2003community,
@@ -112,7 +122,8 @@ year={2022}
[…]
