+++ paper.bib
@@ -1,12 +1,15 @@
-@article{evanega2020coronavirus,
-  title={Coronavirus misinformation: quantifying sources and themes in the COVID-19 ‘infodemic’},
-  author={Evanega, Sarah and Lynas, Mark and Adams, Jordan and Smolenyak, Karinne and Insights, Cision Global},
-  journal={JMIR Preprints},
-  volume={19},
-  number={10},
-  pages={2020},
-  year={2020}
+@article{gallotti2020assessing,
+  title={Assessing the risks of ‘infodemics’ in response to COVID-19 epidemics},
+  author={Gallotti, Riccardo and Valle, Francesco and Castaldo, Nicola and Sacco, Pierluigi and De Domenico, Manlio},
+  journal={Nature human behaviour},
+  volume={4},
+  number={12},
+  pages={1285--1293},
+  year={2020},
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s41562-020-00994-6}
 }
+
 @inproceedings{kephart1993measuring,
   title={Measuring and modeling computer virus prevalence},
   author={Kephart, Jeffrey O and White, Steve R},
