### 2024-07-25 #6796 @majensen
Thanks very much @xianggebenben. The final action before I recommend acceptance is for you to create an archive of your code repository in Zenodo or similar open service. There is an example tutorial at https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729.
Once you do this, can you report the DOI for the archive here? Thanks
