+++ GraphSL/GNN/SLVAE/model.py
@@ -10,11 +10,11 @@ class Encoder(nn.Module):
 
     Attributes:
 
-    - input_dim (int): Dimension of the input.
+        - input_dim (int): Dimension of the input.
 
-    - hidden_dim (int): Dimension of the hidden layer.
+        - hidden_dim (int): Dimension of the hidden layer.
 
-    - latent_dim (int): Dimension of the latent space.
+        - latent_dim (int): Dimension of the latent space.
     """
 
     def __init__(self, input_dim=784, hidden_dim=512, latent_dim=256):
@@ -23,11 +23,11 @@ class Encoder(nn.Module):
 
         Args:
 
-        - input_dim (int): Dimension of the input.
+            - input_dim (int): Dimension of the input.
 
-        - hidden_dim (int): Dimension of the hidden layer.
+            - hidden_dim (int): Dimension of the hidden layer.
 
-        - latent_dim (int): Dimension of the latent space.
+            - latent_dim (int): Dimension of the latent space.
         """
         super(Encoder, self).__init__()
 
@@ -42,13 +42,13 @@ class Encoder(nn.Module):
         Forward pass of the Encoder.
 
         Args:
-        - x (torch.Tensor): Input tensor.
+            - x (torch.Tensor): Input tensor.
 
         Returns:
 
-        - mean (torch.Tensor): The mean of the latent space.
+            - mean (torch.Tensor): The mean of the latent space.
 
-        - log_var (torch.Tensor): The log variance of the latent space.
+            - log_var (torch.Tensor): The log variance of the latent space.
 
         """
         x = self.LeakyReLU(self.linear1(x))
@@ -65,11 +65,11 @@ class Decoder(nn.Module):
 
     Attributes:
 
-    - output_dim (int): Dimension of the output.
+        - output_dim (int): Dimension of the output.
 
-    - hidden_dim (int): Dimension of the hidden layer.
+        - hidden_dim (int): Dimension of the hidden layer.
 
-    - latent_dim (int): Dimension of the latent space.
+        - latent_dim (int): Dimension of the latent space.
     """
 
     def __init__(self, output_dim=784, hidden_dim=512, latent_dim=256):
@@ -78,12 +78,11 @@ class Decoder(nn.Module):
 
         Args:
 
-        - output_dim (int): Dimension of the output.
+            - output_dim (int): Dimension of the output.
 
+            - hidden_dim (int): Dimension of the hidden layer.
 
-        - hidden_dim (int): Dimension of the hidden layer.
-
-        - latent_dim (int): Dimension of the latent space.
+            - latent_dim (int): Dimension of the latent space.
         """
         super(Decoder, self).__init__()
 
@@ -98,11 +97,11 @@ class Decoder(nn.Module):
 
         Args:
 
-        - x (torch.Tensor): Input tensor.
+            - x (torch.Tensor): Input tensor.
 
         Returns:
 
-        - x_hat (torch.Tensor): Decoded output tensor.
+            - x_hat (torch.Tensor): Decoded output tensor.
         """
         x = self.LeakyReLU(self.linear2(x))
         x = self.LeakyReLU(self.linear1(x))
@@ -117,11 +116,11 @@ class VAE(nn.Module):
 
     Attributes:
 
-    - input_dim (int): Dimension of the input.
+        - input_dim (int): Dimension of the input.
 
-    - hidden_dim (int): Dimension of the hidden layer.
+        - hidden_dim (int): Dimension of the hidden layer.
 
-    - latent_dim (int): Dimension of the latent space.
+        - latent_dim (int): Dimension of the latent space.
     """
 
     def __init__(self, input_dim=1, hidden_dim=512, latent_dim=256):
[…]

[…]
