+++ build/lib/GraphSL/Prescribed.py
@@ -3,21 +3,22 @@ import networkx as nx
 import copy
 import torch
 from sklearn.metrics import roc_auc_score, f1_score, accuracy_score, precision_score, recall_score
-from scipy.sparse import csgraph, coo_matrix
-from GraphSL.Evaluation import Metric
-
+from GraphSL.utils import Metric
+from scipy.sparse.csgraph import laplacian as csgraph_laplacian
 
 class LPSI:
     """
     Implement the Label Propagation based Source Identification (LPSI) algorithm.
 
-    Wang, Zheng, et al. "Multiple source detection without knowing the underlying propagation model." Proceedings of the AAAI Conference on Artificial Intelligence. Vol. 31. No. 1. 2017.
+    Wang, Zheng, et al. "Multiple source detection without knowing the underlying propagation model." 
+    Proceedings of the AAAI Conference on Artificial Intelligence. Vol. 31. No. 1. 2017.
     """
 
     def __init__(self):
         """
         Initialize the LPSI module.
         """
+        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
 
     def predict(self, laplacian, num_node, alpha, diff_vec):
         """
@@ -25,118 +26,93 @@ class LPSI:
 
         Args:
 
-        - laplacian (numpy.ndarray): The Laplacian matrix of the graph.
+            - laplacian (torch.Tensor): The Laplacian matrix of the graph.
 
-        - num_node (int): Number of nodes in the graph.
+            - num_node (int): Number of nodes in the graph.
 
-        - alpha (float): the fraction of label information that a node gets from its neighbors (between 0 and 1).
+            - alpha (float): The fraction of label information that a node gets from its neighbors (between 0 and 1).
 
-        - diff_vec (numpy.ndarray): The diffusion vector.
+            - diff_vec (torch.Tensor): The diffusion vector.
 
         Returns:
 
-        - x (numpy.ndarray): Prediction of source nodes.
+            - x (torch.Tensor): Prediction of source nodes.
         """
-        x = (1 - alpha) * np.matmul(np.linalg.pinv(np.eye(N=num_node) -
-                                                   alpha * laplacian, hermitian=True), diff_vec)
+        I = torch.eye(num_node, device=self.device)
+        inv_matrix = torch.linalg.inv(I - alpha * laplacian)
+        x = (1 - alpha) * inv_matrix @ diff_vec
         return x
 
-    def train(
-        self, adj, train_dataset, alpha_list=[
-            0.001, 0.01, 0.1], num_thres=10):
+    def train(self, adj, train_dataset, alpha_list=[0.001, 0.01, 0.1], num_thres=10):
         """
-         Train the LPSI algorithm.
+        Train the LPSI algorithm.
 
         Args:
 
-        - adj (scipy.sparse.csr_matrix): The adjacency matrix of the graph.
+            - adj (scipy.sparse.csr_matrix): The adjacency matrix of the graph.
 
-        - train_dataset (torch.utils.data.dataset.Subset): the training dataset (number of simulations * number of graph nodes * 2 (the first column is seed vector and the second column is diffusion vector)).
+            - train_dataset (torch.utils.data.dataset.Subset): The training dataset.
 
-        - alpha_list (list): List of the fraction of label information that a node gets from its neighbors (between 0 and 1) to try.
+            - alpha_list (list): List of alpha values to try.
 
-        - num_thres (list): Number of threshold values to try.
+            - num_thres (int): Number of threshold values to try.
 
         Returns:
 
[…]
