+++ build/lib/GraphSL/GNN/IVGD/__init__.py
@@ -1,7 +1,4 @@
 from .main import *
 from .correction import *
-from .earlystopping import *
-from .i_deepis import *
-from .preprocessing import *
-from .training import *
 from .validity_net import *
+from .diffusion_model import *
