### 2024-07-29 #6796 @majensen
Thanks @xianggebenben - JOSS prefers that the author line in the archive matches the authors in the paper. Would you be able to add your co-author to the Authors field?
