### 2024-07-29 #6796 @xianggebenben
> Thanks @xianggebenben - JOSS prefers that the author line in the archive matches the authors in the paper. Would you be able to add your co-author to the Authors field?

Got it. I have added my coauthor. Please let me know if you have any questions. Thank you so much for accepting our paper. Feedback from you and reviewers improves our work significantly. I appreciate your help and wish you all the best.
