### 2024-07-29 #6796 @crvernon
:wave: @xianggebenben - Please correct the following before I move to accept this for publication:

In your paper:
- [ ] LINE 7:  The first letter of the sentence "g" should be capitalized.
- [ ] LINE 12:  "Github" should have a capital "H" to appear as "GitHub"
- [ ] LINE 13:  "Github" should have a capital "H" to appear as "GitHub"
- [ ] LINE 23:  Missing space between "viruses" and "(Kephart &..."
- [ ] For your references, ensure that capitalization is maintained where necessary.  For example in LINE 97, the "p" in "python" should be capitalized.  You can do this by placing the characters you wish to preserve formatting for in curly brackets in your bib file. Please check all of your references to ensure these are accounted for.

No additional release is necessary.
