+++ paper.bib
@@ -85,6 +85,7 @@ doi={10.1145/3485447.3512155}
 
 @article{McCabe2022joss,
   doi = {10.21105/joss.04894},
+  title = {cosasi: Graph Diffusion Source Inference in Python},
   url = {https://doi.org/10.21105/joss.04894},
   year = {2022},
   publisher = {The Open Journal},
@@ -92,7 +93,7 @@ doi={10.1145/3485447.3512155}
   number = {80},
   pages = {4894},
   author = {Lucas H. McCabe},
-  title = {cosasi: Graph Diffusion Source Inference in Python}, journal = {Journal of Open Source Software}
+  , journal = {Journal of Open Source Software}
 }
 
 @article{frkaszczak2022rpasdt,
