+++ paper.bib
@@ -1,5 +1,5 @@
 @article{gallotti2020assessing,
-  title={Assessing the risks of ‘infodemics’ in response to COVID-19 epidemics},
+  title={Assessing the risks of ‘infodemics’ in response to {COVID-19} epidemics},
   author={Gallotti, Riccardo and Valle, Francesco and Castaldo, Nicola and Sacco, Pierluigi and De Domenico, Manlio},
   journal={Nature human behaviour},
   volume={4},
@@ -49,7 +49,7 @@ doi={10.1145/3485447.3512155}
   doi={10.1609/aaai.v31i1.10477}
 }
 @inproceedings{zhu2017catch,
-  title={Catch’em all: Locating multiple diffusion sources in networks with partial observations},
+  title={Catch’em all: {Locating} multiple diffusion sources in networks with partial observations},
   author={Zhu, Kai and Chen, Zhen and Ying, Lei},
   booktitle={Proceedings of the AAAI Conference on Artificial Intelligence},
   volume={31},
@@ -58,7 +58,7 @@ doi={10.1145/3485447.3512155}
   doi={10.1609/aaai.v31i1.10746}
 }
 @inproceedings{prakash2012spotting,
-  title={Spotting culprits in epidemics: How many and which ones?},
+  title={Spotting culprits in epidemics: {How} many and which ones?},
   author={Prakash, B Aditya and Vreeken, Jilles and Faloutsos, Christos},
   booktitle={2012 IEEE 12th international conference on data mining},
   pages={11--20},
@@ -97,7 +97,7 @@ doi={10.1145/3485447.3512155}
 }
 
 @article{frkaszczak2022rpasdt,
-  title={RPaSDT—rumor propagation and source detection Toolkit},
+  title={RPaSDT—rumor propagation and source detection {Toolkit}},
   author={Fraszczak, Damian},
   journal={SoftwareX},
   volume={17},
@@ -108,7 +108,7 @@ doi={10.1145/3485447.3512155}
 }
 
 @article{lusseau2003bottlenose,
-  title={The bottlenose dolphin community of doubtful sound features a large proportion of long-lasting associations: can geographic isolation explain this unique trait?},
+  title={The bottlenose dolphin community of {Doubtful} {Sound} features a large proportion of long-lasting associations: {Can} geographic isolation explain this unique trait?},
   author={Lusseau, David and Schneider, Karsten and Boisseau, Oliver J and Haase, Patti and Slooten, Elisabeth and Dawson, Steve M},
   journal={Behavioral Ecology and Sociobiology},
   volume={54},
@@ -154,7 +154,7 @@ doi={10.1145/3485447.3512155}
 }
 
 @article{watts1998collective,
-  title={Collective dynamics of ‘small-world’networks},
+  title={Collective dynamics of ‘small-world’ networks},
   author={Watts, Duncan J and Strogatz, Steven H},
   journal={nature},
   volume={393},
@@ -165,29 +165,8 @@ doi={10.1145/3485447.3512155}
   doi={10.1038/30918}
 }
 
-@inproceedings{leskovec2009meme,
-  title={Meme-tracking and the dynamics of the news cycle},
-  author={Leskovec, Jure and Backstrom, Lars and Kleinberg, Jon},
-  booktitle={Proceedings of the 15th ACM SIGKDD international conference on Knowledge discovery and data mining},
-  pages={497--506},
-  year={2009},
-  doi={10.1145/1557019.1557077}
-}
-
-@article{hogg2012social,
-  title={Social dynamics of digg},
-  author={Hogg, Tad and Lerman, Kristina},
-  journal={EPJ Data Science},
-  volume={1},
-  number={1},
-  pages={1--26},
-  year={2012},
[…]
