### 2024-07-30 #6796 @crvernon
🥳 Congratulations on your new publication @xianggebenben! Many thanks to @majensen for editing and @range-et, @mbeyss, and @rkdan  for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
